chore: 아키텍쳐에 따라 패키지 구조 변경
