+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/dto/CouponResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.coupon.dto;
+package kr.hhplus.be.server.interfaces.coupon.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
