+++ src/main/java/kr/hhplus/be/server/interfaces/order/dto/OrderResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order.dto;
+package kr.hhplus.be.server.interfaces.order.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
